Does not work in BGs due
In mercenary BGs there is no way (as far as I know) to determine a player's actual race, as their race might be Undead for example, when their real race is Human. The addon will think they are Undead, and then will alert in chat the next time the player's real race is encountered. This is undesirable, so I  disabled auto-scanning new players in BGs.